Server/CPP_Server/requirements.txt
absl-py==0.8.1
aniso8601==8.0.0
asn1crypto==1.2.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
boto==2.49.0
boto3==1.10.19
botocore==1.13.19
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
crypto==1.4.1
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
Flask-RESTful==0.3.7
flask-restplus==0.13.0
Flask-SocketIO==4.2.1
gast==0.2.2
gensim==3.8.1
google-pasta==0.1.7
googletrans==2.4.0
grpcio==1.16.1
h5py==2.9.0
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
jieba3k==0.35.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
JPype1==0.7.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-c-kernel==1.2.2
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
konlpy==0.5.1
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
Naked==0.1.31
nbconvert==5.6.1
nbformat==4.4.0
newspaper3k==0.2.8
nltk==3.4.5
notebook==6.0.1
numpy==1.17.3
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.9.2
pycparser==2.19
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.4
pyreadline==2.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-engineio==3.10.0
python-socketio==4.3.1
pytz==2019.3
pywin32==223
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
requests-file==1.4.3
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
shellescape==3.4.1
six==1.12.0
smart-open==1.9.0
soupsieve==1.9.5
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tinysegmenter==0.3
tldextract==2.2.2
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
zipp==0.6.0
